+++ Plugins/TRStatsPlugin/TRStatsPlugin.cs
@@ -294,7 +294,7 @@ namespace TRStats
                 hTitle.transform.SetParent(header.transform, false);
                 Text ht = hTitle.AddComponent<Text>();
                 ht.font = Resources.GetBuiltinResource<Font>("Arial.ttf");
-                ht.text = "TR STATS 1.3.0 (F4)";
+                ht.text = "TR CHEATS 1.3.0 (F4)";
                 ht.alignment = TextAnchor.MiddleCenter;
                 ht.color = new Color(1f, 0.8f, 0.4f);
                 ht.fontSize = 14;